Changes to reflect review feedback
